Research Project on Parallelized Ray Tracing
This research project aims to compare how much
speedup is faced by computing the ray trace of
a scene when using different types of processing elements.
The elements that are compared are cpu and GPU,
where on cpu we compare serial, parallel omp,
and parallel Open MPI implementation of ray tracing,
and on GPU we implement a CUDA ray tracer.